per-project vendoring of self-contained dependencies (fixes #1741) (#1756)
The `codec` project is a dependency for several other projects we plan to open source, and thus must be open sourced. When it is, the open source repo will need to included the vendored yaml libraries that it encapsulates.

My claim is that it's both necessary and correct to move these vendored libs into a `vendor` directory within the `codec` project. Otherwise we get into having to do complex slicing of the top-level vendor directory for OSS repos, and I think this will lead to madness.

Doing this potentially opens us to diamond dependencies, but I think as long as we follow this rule:

- only locally vendor a project if your use of the project is completely self contained; no reference to the locally vendored project can escape the library.

it's totally fine to have multiple versions of the same library vendored in several projects; golang will always try to find the closest sibling to the code its compiling. As long as the vendored code doesn't escape the library vendoring it, no diamond dependency can occur.

To make sure this worked, I did it with two other projects: `cli` and `tbncollect/plugins/consul`. Note that the latter is a sub-project, which I needed to make sure works properly.

In order to get govendor to ignore the TBN_HOME vendor dir, I had to do a little dance to copy the project out into the GO_PATH, run govendor, and then move back in. I tried many other package management approaches, this was better because (1) it's what we're using already and (2) it actually seems to work right.

fbshipit-source-id: 8b6fae4